Refactor additional ingredient fields into reusable component
